requirements.txt
discord.py
requests
asyncio
flask